Restructure simualtors in src directory
* Move controller headers to src/controllers
* Move base State header to src/simulator
* move src/simualtors/open_pulse to src/open_pulse
* Update changelog for GPU backends
