+++ CMakeLists.txt
@@ -276,7 +276,7 @@ if(SKBUILD) # Terra Addon build
 	add_subdirectory(qiskit/providers/aer/pulse/cy)
 	add_subdirectory(qiskit/providers/aer/pulse/qutip_lite/cy)
 	add_subdirectory(qiskit/providers/aer/backends/wrappers)
-	add_subdirectory(src/simulators/open_pulse)
+	add_subdirectory(src/open_pulse)
 else() # Standalone build
 	set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
 	if(CUDA_FOUND)
