+++ qiskit/providers/aer/backends/wrappers/bindings.cc
@@ -6,10 +6,10 @@
 #include "framework/types.hpp"
 #include "framework/pybind_json.hpp"
 
-#include "simulators/qasm/qasm_controller.hpp"
-#include "simulators/statevector/statevector_controller.hpp"
-#include "simulators/unitary/unitary_controller.hpp"
-#include "simulators/controller_execute.hpp"
+#include "controllers/qasm_controller.hpp"
+#include "controllers/statevector_controller.hpp"
+#include "controllers/unitary_controller.hpp"
+#include "controllers/controller_execute.hpp"
 
 PYBIND11_MODULE(controller_wrappers, m) {
     m.def("qasm_controller_execute_json", &AER::controller_execute_json<AER::Simulator::QasmController>, "instance of controller_execute for QasmController");
