+++ src/simulators/density_matrix/densitymatrix_state.hpp
@@ -21,7 +21,7 @@
 
 #include "framework/utils.hpp"
 #include "framework/json.hpp"
-#include "base/state.hpp"
+#include "simulators/state.hpp"
 #include "densitymatrix.hpp"
 #ifdef AER_THRUST_SUPPORTED
 #include "densitymatrix_thrust.hpp"
