+++ src/simulators/extended_stabilizer/extended_stabilizer_state.hpp
@@ -18,7 +18,7 @@
 #include <complex>
 #include <vector>
 
-#include "base/state.hpp"
+#include "simulators/state.hpp"
 #include "framework/json.hpp"
 #include "framework/types.hpp"
