+++ src/simulators/stabilizer/stabilizer_state.hpp
@@ -17,7 +17,7 @@
 
 #include "framework/utils.hpp"
 #include "framework/json.hpp"
-#include "base/state.hpp"
+#include "simulators/state.hpp"
 #include "clifford.hpp"
 
 namespace AER {
