+++ src/simulators/unitary/unitary_state.hpp
@@ -19,7 +19,7 @@
 #define _USE_MATH_DEFINES
 #include <math.h>
 
-#include "base/state.hpp"
+#include "simulators/state.hpp"
 #include "framework/json.hpp"
 #include "framework/utils.hpp"
 #include "unitarymatrix.hpp"
